+++ src/layout/index.vue
@@ -1,5 +1,5 @@
 <template>
-  <PotLayout footer trigger="top" :header-mix="true">
+  <PotLayout footer trigger="top" :header-mix="false">
     <template #default><router-view /></template>
 
     <template #header><LayoutHeader /></template>
@@ -14,6 +14,7 @@
     <template #sidebar="{ collapsed }"><LayoutSidebar :collapsed="collapsed.value" /></template>
 
     <template #logo>Logo Slot</template>
+    <!--<template #trigger>Trigger</template>-->
   </PotLayout>
 </template>