+++ packages/vue-layout/CHANGELOG.md
@@ -1,3 +1,27 @@
+## [0.0.1-beta.3](https://github.com/potjs/pot/compare/vue-layout@0.0.1-beta.2...vue-layout@0.0.1-beta.3) (2021-11-11)
+
+
+### Bug Fixes
+
+* fix trigger to layout ([3e0dfc4](https://github.com/potjs/pot/commit/3e0dfc4c0b0bb851d85912f6a2a0894c92fcd3b6))
+
+
+
+## [0.0.1-beta.2](https://github.com/potjs/pot/compare/vue-layout@0.0.1-beta.1...vue-layout@0.0.1-beta.2) (2021-11-11)
+
+
+### Bug Fixes
+
+* clean code on layout ([845ff3b](https://github.com/potjs/pot/commit/845ff3bb14c3805f4c2b7f1cc49be9db7e064a56))
+* small fix ([d9c8591](https://github.com/potjs/pot/commit/d9c859116a103ff43b2fc158548d0060dee5de12))
+* update ajs to pot ([c345877](https://github.com/potjs/pot/commit/c345877a024b9b620b830e08e026712195f0c0d7))
+* update ajs to pot ([c466d0e](https://github.com/potjs/pot/commit/c466d0e4124943e7d17037f09ae8b99cdaa932a2))
+* update README ([fc294c0](https://github.com/potjs/pot/commit/fc294c06d4ceddb4a95dc8beab456c73619bd93f))
+* update README ([d6ddf80](https://github.com/potjs/pot/commit/d6ddf80f77eeff66f783acac0a9d8b0e9a934cba))
+* update README ([c5f4c90](https://github.com/potjs/pot/commit/c5f4c90750e1c13b35c13705b9da58e0451e637f))
+
+
+
 ## [0.0.1-beta.1](https://github.com/potjs/pot/compare/vue-layout@0.0.1-beta.0...vue-layout@0.0.1-beta.1) (2021-11-06)